requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy==1.23.4
joblib==1.3.2
pydantic==2.5.2
scikit-learn==1.2.2
python-multipart==0.0.6
httpx==0.25.2
python-dotenv==1.0.0
tensorflow-cpu==2.18.0
h5py==3.10.0